Refs #17049: Implement metadata for enums
