+++ rudder-lang/src/ast.rs
@@ -2,6 +2,7 @@
 // SPDX-FileCopyrightText: 2019-2020 Normation SAS
 
 pub mod context;
+pub mod enum_tree;
 pub mod enums;
 pub mod resource;
 pub mod value;
